Parse apmbench results, add ES basicauth
Updates the benchmark parsing to allow to optionally parse the extra
metrics that are published by apmbench. Modified the Elasticsearch
request handling slightly to allow us to authenticate using basicauth
and also work with versions of the stack `8.0.0+`.

Last, created a Go module to allow the project to work with newer Go
versions.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>